fixed i1 and i2 data type (made uint16_t)

+++ v1.0/sketch_esp_WiFi_UART_Bridge.ino
@@ -51,10 +51,10 @@ WiFiClient client;
 
 
 uint8_t buf1[1024];
-uint8_t i1=0;
+uint16_t i1=0;
 
 uint8_t buf2[1024];
-uint8_t i2=0;
+uint16_t i2=0;
